refactor: Moved string converter to its own script and namespace
